refactor and bug fixes after new plugins
